Moved files out of model folder and into respective fragment files for use with adapters.
