+++ app/src/main/java/com/example/musclenerds/ui/exerciseview/Adapters/MuscleGroupAdapter.java
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.musclenerds.R;
-import com.example.musclenerds.model.MuscleGroup;
+import com.example.musclenerds.ui.exerciseview.MuscleGroup;
 
 import java.util.ArrayList;
